Moved llvm_binary_utils.hpp contents to llvm_binary_stubs.hpp
